refactor(builder): convert builder.py into builder/ package
- Split the monolithic `src/pyconfigre/builder.py` into a modular package:
  - `src/pyconfigre/builder/__init__.py` — public re-exports
  - `src/pyconfigre/builder/_merge.py` — `_deep_merge()` utility
  - `src/pyconfigre/builder/raw_builder.py` — `RawConfigBuilder`
  - `src/pyconfigre/builder/config_builder.py` — `ConfigBuilder`
- Preserve `from pyconfigre import ...` and `from pyconfigre.builder import ...` import paths via `__init__.py` re-exports
- Update `__version__` fallback to `0.3.0.dev`
- Mirror the new source layout in the test suite:
  - Create `tests/unit/builder/` package
  - Split `tests/unit/test_builder.py` into `test_raw_builder.py`, `test_config_builder.py`, and `test_merge.py`
- Update mock paths to refer to the new file locations
